backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
python-dotenv==1.0.1
groq==0.9.0
pymupdf==1.24.7
python-docx==1.1.2
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4